+++ windows/flutter/generated_plugins.cmake
@@ -6,6 +6,10 @@ list(APPEND FLUTTER_PLUGIN_LIST
   file_selector_windows
   flutter_secure_storage_windows
   local_auth_windows
+  media_kit_libs_windows_video
+  media_kit_video
+  syncfusion_pdfviewer_windows
+  url_launcher_windows
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST